Update components imports
