+++ src/pages/MyAssignments/index.tsx
@@ -1,23 +1,23 @@
 import { useEffect } from "react";
 import CreateMyAssignmentContainer from "../../components/Lab/Assignment/CreateAssignment/CreateMyAssignmentContainer";
 import ListMyAssignment from "../../components/Lab/Assignment/ListMyAssignment";
-import Terminal from "../../components/Terminal";
+import Terminal from "../../components/Terminal/Terminal";
 import PageLayout from "../../layouts/PageLayout";
 
 type Props = {};
 
 export default function MyAssignments({}: Props) {
-  useEffect(() => {
-    document.title = "ACT Labs | My Assignments";
-  }, []);
+	useEffect(() => {
+		document.title = "ACT Labs | My Assignments";
+	}, []);
 
-  return (
-    <PageLayout heading="My Readiness Lab Assignments">
-      <div className="flex flex-col gap-4">
-        <CreateMyAssignmentContainer />
-        <ListMyAssignment />
-      </div>
-      <Terminal />
-    </PageLayout>
-  );
+	return (
+		<PageLayout heading="My Readiness Lab Assignments">
+			<div className="flex flex-col gap-4">
+				<CreateMyAssignmentContainer />
+				<ListMyAssignment />
+			</div>
+			<Terminal />
+		</PageLayout>
+	);
 }
